[DEV-3387] refactor prefix filtering
