+++ src/config/config.ts
@@ -12,6 +12,7 @@ import { Fiat } from 'src/shared/models/fiat/fiat.entity';
 import { Process } from 'src/shared/services/process.service';
 import { PaymentStandard } from 'src/subdomains/core/payment-link/enums';
 import { KycFile } from 'src/subdomains/generic/kyc/dto/kyc-file.dto';
+import { ContentType } from 'src/subdomains/generic/kyc/enums/content-type.enum';
 import { MailOptions } from 'src/subdomains/supporting/notification/services/mail.service';
 
 export enum Environment {
@@ -568,20 +569,21 @@ export class Configuration {
   downloadTargets = [
     {
       folderName: '02_Identifikationsdokument',
-      filter: (file: KycFile, userDataId: number) =>
-        file.name.endsWith('.pdf') &&
-        (file.name.startsWith(`user/${userDataId}/Identification`) ||
-          file.name.startsWith(`spider/${userDataId}/online-identification`) ||
-          file.name.startsWith(`spider/${userDataId}/video-identification`)),
+      fileTypes: [ContentType.PDF],
+      prefixes: (userDataId: number) => [
+        `user/${userDataId}/Identification`,
+        `spider/${userDataId}/online-identification`,
+        `spider/${userDataId}/video-identification`,
+      ],
+      filter: null,
       reduceFilter: (latest: KycFile, current: KycFile) =>
         new Date(latest.updated) > new Date(current.updated) ? latest : current,
     },
     {
       folderName: '04_Identifizierungsformular',
-      filter: (file: KycFile, userDataId: number) =>
-        file.name.endsWith('.pdf') &&
-        file.name.startsWith(`user/${userDataId}/UserNotes`) &&
-        file.name.includes('Identifizierungsformular'),
+      fileTypes: [ContentType.PDF],
+      prefixes: (userDataId: number) => [`user/${userDataId}/UserNotes`],
+      filter: (file: KycFile) => file.name.includes('Identifizierungsformular'),
       reduceFilter: (latest: KycFile, current: KycFile) =>
         new Date(latest.updated) > new Date(current.updated) ? latest : current,
     },
