+++ src/subdomains/generic/kyc/dto/kyc-file.dto.ts
@@ -3,6 +3,7 @@ import { Blob } from 'src/integration/infrastructure/azure-storage.service';
 import { UserData } from '../../user/models/user-data/user-data.entity';
 import { KycWebhookData } from '../../user/services/webhook/dto/kyc-webhook.dto';
 import { KycStep } from '../entities/kyc-step.entity';
+import { ContentType } from '../enums/content-type.enum';
 
 export enum FileType {
   NAME_CHECK = 'NameCheck',
@@ -15,18 +16,6 @@ export enum FileType {
   RESIDENCE_PERMIT = 'ResidencePermit',
   ADDITIONAL_DOCUMENTS = 'AdditionalDocuments',
   AUTHORITY = 'Authority',
-}
-
-export enum ContentType {
-  PNG = 'image/png',
-  JPEG = 'image/jpeg',
-  JPG = 'image/jpg',
-  JSON = 'application/json',
-  PDF = 'application/pdf',
-  TEXT = 'text/plain',
-  XML = 'text/xml',
-  ZIP = 'application/zip',
-  MP3 = 'audio/mpeg',
 }
 
 export enum KycReportType {
