+++ src/subdomains/generic/kyc/enums/content-type.enum.ts
@@ -1,0 +1,11 @@
+export enum ContentType {
+  PNG = 'image/png',
+  JPEG = 'image/jpeg',
+  JPG = 'image/jpg',
+  JSON = 'application/json',
+  PDF = 'application/pdf',
+  TEXT = 'text/plain',
+  XML = 'text/xml',
+  ZIP = 'application/zip',
+  MP3 = 'audio/mpeg',
+}
