+++ src/subdomains/generic/kyc/services/integration/kyc-document.service.ts
@@ -1,8 +1,9 @@
 import { Injectable, UnsupportedMediaTypeException } from '@nestjs/common';
 import { AzureStorageService, BlobContent } from 'src/integration/infrastructure/azure-storage.service';
 import { UserData } from 'src/subdomains/generic/user/models/user-data/user-data.entity';
-import { ContentType, FileType, KycFile } from '../../dto/kyc-file.dto';
+import { FileType, KycFile } from '../../dto/kyc-file.dto';
 import { KycStep } from '../../entities/kyc-step.entity';
+import { ContentType } from '../../enums/content-type.enum';
 import { KycFileService } from '../kyc-file.service';
 
 @Injectable()
@@ -21,11 +22,9 @@ export class KycDocumentService {
     return this.listFilesByPrefix(`spider/${userDataId}${isOrganization ? '-organization' : ''}/`);
   }
 
-  async listSpiderKycFiles(userDataId: number): Promise<KycFile[]> {
-    const onlineIdentFiles = await this.listFilesByPrefix(`spider/${userDataId}/online-identification/`);
-    const videoIdentFiles = await this.listFilesByPrefix(`spider/${userDataId}/video-identification/`);
-
-    return [...onlineIdentFiles, ...videoIdentFiles];
+  async listFilesByPrefixes(prefixes: string[]): Promise<KycFile[]> {
+    const files = await Promise.all(prefixes.map((p) => this.listFilesByPrefix(p)));
+    return files.flat();
   }
 
   async listFilesByPrefix(prefix: string): Promise<KycFile[]> {
