+++ src/subdomains/generic/kyc/services/kyc-client.service.ts
@@ -9,7 +9,8 @@ import { UserService } from '../../user/models/user/user.service';
 import { WalletService } from '../../user/models/wallet/wallet.service';
 import { PaymentWebhookData } from '../../user/services/webhook/dto/payment-webhook.dto';
 import { WebhookDataMapper } from '../../user/services/webhook/mapper/webhook-data.mapper';
-import { ContentType, FileType, KycClientDataDto, KycFile, KycReportDto, KycReportType } from '../dto/kyc-file.dto';
+import { FileType, KycClientDataDto, KycFile, KycReportDto, KycReportType } from '../dto/kyc-file.dto';
+import { ContentType } from '../enums/content-type.enum';
 import { KycDocumentService } from './integration/kyc-document.service';
 
 @Injectable()
