+++ src/subdomains/generic/user/models/kyc/kyc.service.ts
@@ -11,7 +11,8 @@ import { LanguageDtoMapper } from 'src/shared/models/language/dto/language-dto.m
 import { DfxLogger } from 'src/shared/services/dfx-logger';
 import { HttpService } from 'src/shared/services/http.service';
 import { Util } from 'src/shared/utils/util';
-import { ContentType, FileType, KycFile } from 'src/subdomains/generic/kyc/dto/kyc-file.dto';
+import { FileType, KycFile } from 'src/subdomains/generic/kyc/dto/kyc-file.dto';
+import { ContentType } from 'src/subdomains/generic/kyc/enums/content-type.enum';
 import { KycDocumentService } from 'src/subdomains/generic/kyc/services/integration/kyc-document.service';
 import {
   Blank,
