+++ src/subdomains/generic/user/models/user-data/dto/upload-file.dto.ts
@@ -1,5 +1,6 @@
 import { IsEnum, IsNotEmpty, IsNumber, IsString, ValidateIf } from 'class-validator';
-import { ContentType, FileType } from 'src/subdomains/generic/kyc/dto/kyc-file.dto';
+import { FileType } from 'src/subdomains/generic/kyc/dto/kyc-file.dto';
+import { ContentType } from 'src/subdomains/generic/kyc/enums/content-type.enum';
 
 export class UploadFileDto {
   @IsNotEmpty()
