+++ src/subdomains/supporting/support-issue/services/support-issue.service.ts
@@ -7,7 +7,7 @@ import {
 } from '@nestjs/common';
 import { BlobContent } from 'src/integration/infrastructure/azure-storage.service';
 import { Util } from 'src/shared/utils/util';
-import { ContentType } from 'src/subdomains/generic/kyc/dto/kyc-file.dto';
+import { ContentType } from 'src/subdomains/generic/kyc/enums/content-type.enum';
 import { UserDataService } from 'src/subdomains/generic/user/models/user-data/user-data.service';
 import { FindOptionsWhere, In, IsNull, MoreThan, Not } from 'typeorm';
 import { TransactionService } from '../../payment/services/transaction.service';
